pyproject.toml
[tool.poetry]
name = "XApps"
version = "0.1.0"
description = "Android Apk Downloader"
authors = ["Leorio Paradinight <62891774+code-rgb@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.7.4"
beautifulsoup4 = "^4.9.3"
pyppeteer = "^0.2.5"
fake-useragent = "^0.1.11"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
